+++ src/main.tsx
@@ -2,6 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { HashRouter } from 'react-router-dom'
 import { ThemeProvider } from './shared/ThemeContext'
+import { DialogProvider } from './shared/DialogContext'
 import App from './App'
 import './index.css'
 
@@ -9,7 +10,9 @@ createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <HashRouter>
       <ThemeProvider>
-        <App />
+        <DialogProvider>
+          <App />
+        </DialogProvider>
       </ThemeProvider>
     </HashRouter>
   </StrictMode>,